feat: enable universal NNUE WASM by default
