+++ apps/web/.env.example
@@ -2,6 +2,9 @@
 VITE_NNUE_MANIFEST_URL=https://example/manifest.json
 # Optional
 # VITE_DEFAULT_NNUE_PRESET=ramu
+# WASM エンジンの要求スレッド数。未設定時は4。ブラウザ上限でclampされ、
+# SharedArrayBufferが利用できない場合はsingle-threaded版へフォールバックする。
+# VITE_WASM_THREADS=4
 # VITE_BASE_PATH=/
 # VITE_ALLOWED_HOSTS=your-host.example.com,another-host.example.com
 # rshogi 棋譜配信 API のベース URL。
