+++ apps/web/src/pages/rshogi-viewer/RshogiViewerPage.tsx
@@ -1,32 +1,16 @@
-import { createWasmEngineClient } from "@shogi/engine-wasm";
 import type { EngineOption } from "@shogi/ui";
 import { RshogiCsaViewer } from "@shogi/ui";
 import { getRouteApi } from "@tanstack/react-router";
 import type { ReactElement } from "react";
 import { HeaderNav } from "../../components/HeaderNav";
 import { PageHeader } from "../../components/PageHeader";
-
-const resolveWasmThreads = (): number => {
-    const fallback = import.meta.env.DEV ? 4 : 1;
-    const raw = import.meta.env.VITE_WASM_THREADS;
-    if (typeof raw !== "string" || raw.trim() === "") return fallback;
-    const parsed = Number(raw);
-    if (!Number.isFinite(parsed) || parsed < 1) return fallback;
-    return Math.trunc(parsed);
-};
-
-const wasmThreads = resolveWasmThreads();
+import { createWebWasmEngineClient } from "../../platform/wasm-engine-client";
 
 const engineOptions: EngineOption[] = [
     {
         id: "wasm",
         label: "内蔵エンジン",
-        createClient: () =>
-            createWasmEngineClient({
-                stopMode: "terminate",
-                defaultInitOptions: { threads: wasmThreads },
-                logWarningsToConsole: true,
-            }),
+        createClient: createWebWasmEngineClient,
         kind: "internal",
     },
 ];
